# src/swagger-spec/business-api-swagger-spec.json
{"openapi":"3.0.0","paths":{"/aml/upload-document":{"post":{"operationId":"HttpRestController_uploadDocument","summary":"upload-document","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadDocumentReq"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdResponse"}}}},"403":{"description":""}},"tags":["aml"]}},"/aml/update-document/{docId}":{"put":{"operationId":"HttpRestController_updateDocument","summary":"reupload document","parameters":[{"name":"docId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadDocumentReq"}}}},"responses":{"200":{"description":""}},"tags":["aml"]}},"/aml/download-document/{id}":{"get":{"operationId":"HttpRestController_downloadDocument","summary":"download-document by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetKybDocumentsResponse"}}}}},"tags":["aml"]}},"/aml/get-documents/{purpose}/{organization}":{"get":{"operationId":"HttpRestController_getUploadedDocument","summary":"get-documents","parameters":[{"name":"purpose","required":true,"in":"path","schema":{"type":"string"}},{"name":"organization","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetKybDocumentsResponse"}}}}},"tags":["aml"]}},"/aml/documents/{id}/{tenantId}/{organization}":{"get":{"operationId":"HttpRestController_getDocumentVerification","summary":"Get kyb verification response data","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"tenantId","required":true,"in":"path","schema":{"type":"string"}},{"name":"organization","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"502":{"description":""}},"tags":["aml"]}},"/aml/documents/{id}":{"put":{"operationId":"HttpRestController_updateDocumentStatus","summary":"Update kyb document status","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDocumentStatusReq"}}}},"responses":{"200":{"description":""},"502":{"description":""}},"tags":["aml"]}}},"info":{"title":"Business","description":"Business API Microservice","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"/api/v1"}],"components":{"securitySchemes":{"access-token":{"scheme":"Bearer","bearerFormat":"Bearer","description":"Please enter token in following format: Bearer <JWT>","name":"Authorization","type":"http","in":"Header"}},"schemas":{"UploadDocumentReq":{"type":"object","properties":{"organization":{"type":"string","default":1},"organizationName":{"type":"string"},"purpose":{"type":"string","default":"kyb"},"status":{"type":"string","default":"pending"},"documentType":{"type":"string","default":"id_back_side"},"file":{"type":"file","format":"binary"},"metadata":{"type":"string"}},"required":["organization","organizationName","purpose","documentType","metadata"]},"IdResponse":{"type":"object","properties":{"id":{"type":"string","example":"2cdc8ab1-6d50-49cc-ba14-54e4ac7ec231"}},"required":["id"]},"Status":{"type":"object","properties":{"isError":{"type":"boolean"},"code":{"type":"number"},"description":{"type":"string"},"type":{"type":"string","enum":["done","in_progress","error","unknown","canceled","validation_fail","unauthorized"]},"timestamp":{"format":"date-time","type":"string"}},"required":["isError","code","description","type","timestamp"]},"DocumentDto":{"type":"object","properties":{"purpose":{"type":"array","items":{"type":"string"}},"name":{"type":"string"},"documentType":{"type":"string"},"location":{"type":"string"},"status":{"type":"string"},"metadata":{"type":"string"},"verificationResponse":{"type":"object"}},"required":["purpose","name","documentType","location","status","metadata","verificationResponse"]},"GetKybDocumentsResponse":{"type":"object","properties":{}},"UpdateDocumentStatusReq":{"type":"object","properties":{"status":{"type":"string"}},"required":["status"]}}}}